Namespace StudentLoans to Policies::StudentLoans
